refactor: proxy wiring for EventStreamManager
