+++ platform-sdk/swirlds-platform-core/src/main/java/com/swirlds/platform/SwirldsPlatform.java
@@ -94,6 +94,7 @@ import com.swirlds.platform.event.preconsensus.PcesFileTracker;
 import com.swirlds.platform.event.preconsensus.PcesReplayer;
 import com.swirlds.platform.event.preconsensus.PcesSequencer;
 import com.swirlds.platform.event.preconsensus.PcesWriter;
+import com.swirlds.platform.event.stream.DefaultEventStreamManager;
 import com.swirlds.platform.event.stream.EventStreamManager;
 import com.swirlds.platform.event.validation.AddressBookUpdate;
 import com.swirlds.platform.event.validation.DefaultInternalEventValidator;
@@ -536,7 +537,7 @@ public class SwirldsPlatform implements Platform {
             eventStreamManagerName = String.valueOf(selfId);
         }
 
-        final EventStreamManager eventStreamManager = new EventStreamManager(
+        final EventStreamManager defaultEventStreamManager = new DefaultEventStreamManager(
                 platformContext,
                 time,
                 threadManager,
@@ -660,7 +661,7 @@ public class SwirldsPlatform implements Platform {
                 stateSignatureCollector,
                 transactionPrehandler,
                 consensusRoundHandler,
-                eventStreamManager,
+                defaultEventStreamManager,
                 futureEventBuffer,
                 issDetector,
                 issHandler,
