+++ platform-sdk/swirlds-platform-core/src/test/java/com/swirlds/platform/event/stream/EventStreamManagerTest.java
@@ -25,9 +25,14 @@ import com.swirlds.common.crypto.Hash;
 import com.swirlds.common.stream.MultiStream;
 import com.swirlds.common.stream.RunningEventHashUpdate;
 import com.swirlds.common.test.fixtures.RandomUtils;
+import com.swirlds.common.test.fixtures.platform.TestPlatformContextBuilder;
+import com.swirlds.common.wiring.component.ComponentWiring;
+import com.swirlds.common.wiring.model.WiringModel;
+import com.swirlds.common.wiring.schedulers.builders.TaskSchedulerType;
 import com.swirlds.platform.internal.EventImpl;
 import java.util.List;
 import java.util.Random;
+import java.util.concurrent.ForkJoinPool;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.ValueSource;
@@ -35,7 +40,7 @@ import org.junit.jupiter.params.provider.ValueSource;
 class EventStreamManagerTest {
     private static final MultiStream<EventImpl> multiStreamMock = mock(MultiStream.class);
     private static final EventStreamManager eventStreamManager =
-            new EventStreamManager(Time.getCurrent(), multiStreamMock, EventStreamManagerTest::isFreezeEvent);
+            new DefaultEventStreamManager(Time.getCurrent(), multiStreamMock, EventStreamManagerTest::isFreezeEvent);
 
     private static final EventImpl freezeEvent = mock(EventImpl.class);
 
@@ -51,7 +56,18 @@ class EventStreamManagerTest {
             verify(multiStreamMock, never()).close();
         }
 
-        eventStreamManager.addEvents(List.of(freezeEvent));
+        final WiringModel model = WiringModel.create(
+                TestPlatformContextBuilder.create().build(), Time.getCurrent(), ForkJoinPool.commonPool());
+        final ComponentWiring<EventStreamManager, Void> wiring = new ComponentWiring<>(
+                model,
+                EventStreamManager.class,
+                model.schedulerBuilder("eventStreamManager")
+                        .withType(TaskSchedulerType.DIRECT)
+                        .build()
+                        .cast());
+        wiring.bind(eventStreamManager);
+
+        wiring.getInputWire(EventStreamManager::addEvents).inject(List.of(freezeEvent));
         verify(multiStreamMock).addObject(freezeEvent);
         // for freeze event, multiStream should be closed after adding it
         verify(multiStreamMock).close();
@@ -67,7 +83,20 @@ class EventStreamManagerTest {
     void setStartWriteAtCompleteWindowTest(final boolean startWriteAtCompleteWindow) {
         final Random random = RandomUtils.getRandomPrintSeed();
         final Hash runningHash = RandomUtils.randomHash(random);
-        eventStreamManager.updateRunningHash(new RunningEventHashUpdate(runningHash, startWriteAtCompleteWindow));
+
+        final WiringModel model = WiringModel.create(
+                TestPlatformContextBuilder.create().build(), Time.getCurrent(), ForkJoinPool.commonPool());
+        final ComponentWiring<EventStreamManager, Void> wiring = new ComponentWiring<>(
+                model,
+                EventStreamManager.class,
+                model.schedulerBuilder("eventStreamManager")
+                        .withType(TaskSchedulerType.DIRECT)
+                        .build()
+                        .cast());
+        wiring.bind(eventStreamManager);
+
+        wiring.getInputWire(EventStreamManager::updateRunningHash)
+                .inject(new RunningEventHashUpdate(runningHash, startWriteAtCompleteWindow));
         verify(multiStreamMock).setRunningHash(runningHash);
     }
